Tests added for init_binding validation
